backend/requirements.txt
Django>=5.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
drf-spectacular>=0.27
django-filter>=24.2
django-cors-headers>=4.4
Pillow>=10.4
python-dotenv>=1.0
yookassa>=3.0
openai>=1.0
black>=24.0
isort>=5.13
flake8>=7.0
psycopg2-binary>=2.9
gunicorn>=21.2
